requirements.txt
Pillow==4.0.0
google-cloud-vision==0.42.0
google-cloud-storage
